Don't rely on implicit cast to long.
It's important that the subtraction happens after the values has been converted to long, else an unsigned integer overflow can happen.
My compiler produce a binary which subtracts first and thus it worked for me.
EdB's compiler produce a binary which converts first and thus an unsigned integer overflow happened for him.

This patch ensures that all compilers convert to long before they do the subtraction.

This change has been suggested by EdB.

+++ trunk/src/game/time.cpp
@@ -60,7 +60,7 @@ void Time::LetRealTimePassUntilFrameEnd()
   ASSERT(!IsWaiting());
   long delay;
   do {
-    delay = current_time - stopwatch.GetValue();
+    delay = static_cast<long>(current_time) - stopwatch.GetValue();
     if (delay > 0) {
       SDL_Delay((uint)delay);
       MSG_DEBUG("time.skip","Do nothing for: %d", delay);
